Fix document updateAt mapping (#254)
There was an issue with the updatedAt of a document showing another
document updatedAt. Specifically, the updatedAt in the document list
was being reversed and reflecting a different document's value.

This issue occured during the bulk retrieval of document lists using
yorkie-team/yorkie#931. In this process, there was no guarantee that the
order of the keys passed to the DB query matches the order of the
documents in the query result.